API key support and OAuth 2.0 client credentials flow  (#469)
* get client creds by client_id

* allow client_credentials grant

* client API

* client credentials flow WIP

* datetime util

* client token api progress

* no id token when using client credentials grant

* oauth exceptions

* client api fix

* working client_credentials grant

* add missing provider id

* client token list and label support

* remove unused obsolete func

* short expiration of client credentials grant token

* short expiration of client credentials grant token

* clean up client token api

* flake8

* update CHANGELOG.md

* api key request uses tenant instead of scope

* fix defaul condition

* error handling

* client auth error handling

* client auth error handling

* refactor token endpoint authentication

* refactor token endpoint authentication

* flake8

* rename expiration to exp

* lazy documentation for api keys

* enhance with ai

* fix typo

+++ seacatauth/openidconnect/handler/token.py
@@ -94,12 +94,25 @@ class TokenHandler(object):
 		form_data = await request.post()
 		from_ip = generic.get_request_access_ips(request)
 
-		# 3.1.3.2.  Token Request Validation
+		# Authenticate the client
+		try:
+			client = await self.OpenIdConnectService.ClientService.authenticate_client_request(request)
+		except exceptions.ClientAuthenticationError as e:
+			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Unauthorized client.", struct_data={
+				"from_ip": from_ip,
+				"client_id": e.ClientID,
+				"redirect_uri": form_data.get("redirect_uri"),
+			})
+			return self.token_error_response(request, TokenRequestErrorResponseCode.UnauthorizedClient)
+
+		# Choose flow based on grant_type
 		grant_type = form_data.get("grant_type")
 		if grant_type == const.OAuth2.GrantType.AUTHORIZATION_CODE:
-			return await self._authorization_code_grant(request, from_ip)
+			process_token_request = self._authorization_code_grant(request, client, from_ip)
 		elif grant_type == const.OAuth2.GrantType.REFRESH_TOKEN:
-			return await self._refresh_token_grant(request, from_ip)
+			process_token_request = self._refresh_token_grant(request, client, from_ip)
+		elif grant_type == const.OAuth2.GrantType.CLIENT_CREDENTIALS:
+			process_token_request = self._client_credentials_grant(request, client, from_ip)
 		else:
 			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Unsupported grant type.", struct_data={
 				"from_ip": from_ip,
@@ -109,12 +122,44 @@ class TokenHandler(object):
 			})
 			return self.token_error_response(request, TokenRequestErrorResponseCode.UnsupportedGrantType)
 
+		try:
+			return await process_token_request
+		except exceptions.ClientAuthenticationError as e:
+			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Unauthorized client.", struct_data={
+				"from_ip": from_ip,
+				"grant_type": grant_type,
+				"client_id": e.ClientID,
+				"redirect_uri": form_data.get("redirect_uri"),
+			})
+			return self.token_error_response(request, TokenRequestErrorResponseCode.UnauthorizedClient)
+		except exceptions.OAuth2InvalidClient as e:
+			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Invalid client.", struct_data={
+				"from_ip": from_ip,
+				"grant_type": grant_type,
+				"client_id": e.ClientId,
+				"redirect_uri": form_data.get("redirect_uri"),
+			})
+			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidClient)
+		except exceptions.OAuth2InvalidScope as e:
+			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Invalid scope.", struct_data={
+				"from_ip": from_ip,
+				"grant_type": grant_type,
+				"client_id": e.ClientId,
+				"scope": e.Scope,
+				"redirect_uri": form_data.get("redirect_uri"),
+			})
+			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidScope)
 
-	async def _authorization_code_grant(self, request, from_ip):
+
+	async def _authorization_code_grant(
+		self,
+		request: aiohttp.web.Request,
+		client: dict,
+		from_ip: list
+	) -> aiohttp.web.Response:
+
 		form_data = await request.post()
-
-		client_id = await self.OpenIdConnectService.ClientService.authenticate_client_request(
-			request, expected_client_id=None)
+		client_id = client["_id"]
 
 		# Get session by code
 		try:
@@ -146,15 +191,6 @@ class TokenHandler(object):
 			)
 			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidGrant)
 
-		except exceptions.ClientAuthenticationError as e:
-			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Cannot verify client ({}).".format(e), struct_data={
-				"from_ip": from_ip,
-				"grant_type": const.OAuth2.GrantType.AUTHORIZATION_CODE,
-				"client_id": e.ClientID,
-				"redirect_uri": form_data.get("redirect_uri"),
-			})
-			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidClient)
-
 		except exceptions.URLValidationError:
 			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Redirect URI mismatch.", struct_data={
 				"from_ip": from_ip,
@@ -163,6 +199,19 @@ class TokenHandler(object):
 				"redirect_uri": form_data.get("redirect_uri"),
 			})
 			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidRequest)
+
+		except asab.exceptions.ValidationError:
+			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Invalid request.", struct_data={
+				"from_ip": from_ip,
+				"grant_type": const.OAuth2.GrantType.AUTHORIZATION_CODE,
+				"client_id": client_id,
+				"redirect_uri": form_data.get("redirect_uri"),
+			})
+			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidRequest)
+
+		if client_id != session.OAuth2.ClientId:
+			raise exceptions.ClientAuthenticationError(
+				"Client ID in token request does not match the one used in authorization request.")
 
 		# Establish and propagate track ID
 		session = await self.set_track_id(request, session, from_ip)
@@ -199,11 +248,15 @@ class TokenHandler(object):
 		return asab.web.rest.json_response(request, response_payload, headers=headers)
 
 
-	async def _refresh_token_grant(self, request, from_ip):
-		form_data = await request.post()
+	async def _refresh_token_grant(
+		self,
+		request: aiohttp.web.Request,
+		client: dict,
+		from_ip: list
+	) -> aiohttp.web.Response:
 
-		client_id = await self.OpenIdConnectService.ClientService.authenticate_client_request(
-			request, expected_client_id=None)
+		form_data = await request.post()
+		client_id = client["_id"]
 
 		# Get session by refresh token
 		try:
@@ -232,6 +285,10 @@ class TokenHandler(object):
 		# Refresh is not supported for algorithmic sessions (yet)
 		assert not session.is_algorithmic()
 
+		if client_id != session.OAuth2.ClientId:
+			raise exceptions.ClientAuthenticationError(
+				"Client ID in token request does not match the one used in authorization request.")
+
 		# Delete the used refresh token and the current access token
 		await self.SessionService.TokenService.delete_tokens_by_session_id(session.SessionId)
 
@@ -248,6 +305,91 @@ class TokenHandler(object):
 		scope = form_data.get("scope")
 
 		response_payload = await self._refresh_session_and_issue_tokens(session, scope=scope)
+
+		headers = {
+			"Cache-Control": "no-store",
+			"Pragma": "no-cache",
+		}
+
+		return asab.web.rest.json_response(request, response_payload, headers=headers)
+
+
+	async def _client_credentials_grant(
+		self,
+		request: aiohttp.web.Request,
+		client: dict,
+		from_ip: list
+	) -> aiohttp.web.Response:
+
+		form_data = await request.post()
+		client_id = client["_id"]
+
+		if "scope" not in form_data:
+			AuditLogger.log(
+				asab.LOG_NOTICE,
+				"Token request denied: Missing scope parameter.",
+				struct_data={
+					"from_ip": from_ip,
+					"grant_type": const.OAuth2.GrantType.CLIENT_CREDENTIALS,
+					"client_id": form_data.get("client_id"),
+				}
+			)
+			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidRequest)
+
+		scope = form_data["scope"].split(" ")
+
+		try:
+			tokens = await self.OpenIdConnectService.issue_token_for_client_credentials(
+				client_id=client_id,
+				scope=scope,
+			)
+
+		except exceptions.CredentialsNotFoundError:
+			AuditLogger.log(
+				asab.LOG_NOTICE,
+				"Token request denied: Client does not have Seacat Auth credentials enabled.",
+				struct_data={
+					"from_ip": from_ip,
+					"grant_type": const.OAuth2.GrantType.CLIENT_CREDENTIALS,
+					"client_id": client_id,
+				}
+			)
+			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidClient)
+
+		except exceptions.NoTenantsError:
+			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Client has no tenants.", struct_data={
+				"from_ip": from_ip,
+				"grant_type": const.OAuth2.GrantType.CLIENT_CREDENTIALS,
+				"client_id": client_id,
+				"scope": " ".join(scope),
+			})
+			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidScope)
+
+		except exceptions.AccessDeniedError:
+			AuditLogger.log(asab.LOG_NOTICE, "Token request denied: Unauthorized tenant access.", struct_data={
+				"from_ip": from_ip,
+				"grant_type": const.OAuth2.GrantType.CLIENT_CREDENTIALS,
+				"client_id": client_id,
+				"scope": " ".join(scope),
+			})
+			return self.token_error_response(request, TokenRequestErrorResponseCode.InvalidScope)
+
+		# Token request successful
+		session = tokens["session"]
+		AuditLogger.log(asab.LOG_NOTICE, "Token request granted.", struct_data={
+			"cid": session.Credentials.Id,
+			"sid": session.SessionId,
+			"client_id": client_id,
+			"grant_type": const.OAuth2.GrantType.CLIENT_CREDENTIALS,
+			"from_ip": from_ip,
+		})
+
+		response_payload = {
+			"token_type": "Bearer",
+			"scope": " ".join(session.OAuth2.Scope),
+			"access_token": tokens["access_token"],
+			"expires_in": int((session.Session.Expiration - datetime.datetime.now(datetime.UTC)).total_seconds()),
+		}
 
 		headers = {
 			"Cache-Control": "no-store",
@@ -308,15 +450,6 @@ class TokenHandler(object):
 
 		# TODO: If possible, verify that the Authorization Code has not been previously used.
 
-		# Verify client credentials if required
-		client_id = await self.OpenIdConnectService.ClientService.authenticate_client_request(
-			request, expected_client_id=session.OAuth2.ClientId)
-
-		# Ensure the Authorization Code was issued to the authenticated Client
-		if client_id != session.OAuth2.ClientId:
-			raise exceptions.ClientAuthenticationError(
-				"Client ID in token request does not match the one used in authorization request.")
-
 		if self.ValidateRedirectUri:
 			# Ensure that the redirect_uri parameter value is identical to the redirect_uri parameter value
 			# that was included in the initial Authorization Request.
@@ -344,15 +477,6 @@ class TokenHandler(object):
 		session = await self.OpenIdConnectService.get_session_by_refresh_token(refresh_token)
 
 		# TODO: If possible, verify that the Refresh Token has not been previously used.
-
-		# Verify client credentials if required
-		client_id = await self.OpenIdConnectService.ClientService.authenticate_client_request(
-			request, expected_client_id=session.OAuth2.ClientId)
-
-		# Ensure the Authorization Code was issued to the authenticated Client
-		if client_id != session.OAuth2.ClientId:
-			raise exceptions.ClientAuthenticationError(
-				"Client ID in token request does not match the one used in authorization request.")
 
 		return session
